It compiles! Save point.
I'm starting to understand lifetimes better, and be aware of when I mean to move
something and when I mean to borrow it. This creates an ExecutionResult struct
that knows how to log itself, experiments with different ways of passing around
pairs of things, and how to consume and move the values in those pairs.

I actually have no idea if this works, I haven't run it. I also think the
request tests might have been failing because I set up a duplicate mock right in
the execute() method, which may have been stomping on the ones set up in the
tests. We'll see.